Corrected some typos and moved some methods from console to warehouse.
