+++ src/classes/Warehouse.java
@@ -1,6 +1,6 @@
 package classes;
 
-import functions.Consolle;
+import functions.Console;
 
 import java.util.*;
 
@@ -146,7 +146,7 @@ public class Warehouse {
                     scnRemoveCart();
                     break;
                 case 5:
-                    Consolle.scnResearchSeller(stock);
+                    Console.scnResearchSeller(stock);
                     break;
             }
         } while (!input.equalsIgnoreCase("q"));
@@ -181,7 +181,7 @@ public class Warehouse {
         for (int i = 0; i < cartsName.size(); i++) {
             toPrint = toPrint.concat(i + ". " + cartsName.get(i) + " ");
         }
-        int intChoice = -1;
+        int intChoice;
         do {
             System.out.print("Seleziona il carrello che vorresti eliminare:\n" +
                     toPrint + "\n" +
@@ -200,6 +200,105 @@ public class Warehouse {
 
         carts.remove(cartsName.get(intChoice));
     }
+    public static void checkIfEmptyMap(Map<Product, Integer> originalMap) {
+        if (originalMap.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        try {
+            printFilteredElements(originalMap);
+        } catch (IllegalArgumentException e) {
+            System.out.println("Non è stato trovato nessun elemento corrispondente.\n");
+        }
+    }
 
+    public static boolean checkIfPriceIsInRange(double maxLimit, double minLimit, double checkPrice) {
+        return minLimit <= checkPrice && checkPrice <= maxLimit;
+    }
+
+    public static void printFilteredElements(Map<Product, Integer> map) {  //stampa risultato ricerca
+        for (Product item : map.keySet()) {
+            System.out.println(item.toStringExtended() + "\nQuantity: " + map.get(item) + "\n \n");
+        }
+    }
+
+    public static Map<Product, Integer> researchKindOfProduct(Map<Product, Integer> warehouse, String input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getArticle().equalsIgnoreCase(input)) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchProducer(Map<Product, Integer> warehouse, String input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getProducer().equalsIgnoreCase(input)) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchModel(Map<Product, Integer> warehouse, String input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getModel().equalsIgnoreCase(input)) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchID(Map<Product, Integer> warehouse, int input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getId() == input) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchSellPrice(Map<Product, Integer> warehouse, double input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getSellPrice() == input) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchPurchasePrice(Map<Product, Integer> warehouse, double input) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (item.getPurchasePrice() == input) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchPurchasePriceInRange(Map<Product, Integer> warehouse, double minLimit, double maxLimit) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (checkIfPriceIsInRange(maxLimit, minLimit, item.getPurchasePrice())) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
+
+    public static Map<Product, Integer> researchSellPriceInRange(Map<Product, Integer> warehouse, double minLimit, double maxLimit) {
+        Map<Product, Integer> filteredMap = new HashMap<>();
+        for (Product item : warehouse.keySet()) {
+            if (checkIfPriceIsInRange(maxLimit, minLimit, item.getSellPrice())) {
+                filteredMap.put(item, warehouse.get(item));
+            }
+        }
+        return filteredMap;
+    }
 }
